% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_required_columns}
\alias{check_required_columns}
\title{Check Required Columns}
\usage{
check_required_columns(data, required)
}
\arguments{
\item{data}{Data frame}

\item{required}{Character vector: required column names}
}
\description{
Helper function to check if required columns exist in data.
}
\keyword{internal}
