% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_indicator_params}
\alias{validate_indicator_params}
\title{Validate Indicator Parameters}
\usage{
validate_indicator_params(n, data, min_n = 1)
}
\arguments{
\item{n}{Integer: period parameter}

\item{data}{Data frame}

\item{min_n}{Integer: minimum allowed value for n}
}
\description{
Helper function to validate common indicator parameters.
}
\keyword{internal}
